Add poetry as the dependency manager
Adding Poetry in order to manage packages and update packages.
Chose poetry because it is the management tool I am more
comfortable with and has less political contention compared
to pyenv or conda. This commit adds the toml file as well as
the lock file of the packages the repo will be using.