Migrate to Quart
In the current moment ASGI has been made possible. In
addition to being in a world of python3, we can now
be truly multi-threaded. There are a few options some
of the leading ones are FastApi, Django3, Quart. Since
I was using Flask it is easier to migrate to something
that keeps the same configuration. Quart accomplishes
the transition to ASGI without changing much of the
structure. This commit introduces the migration to
Quart as the web framework. This commit introduces ASGI.